Vertex: Properly center using expand = "none"
Previous behaviour would not center the text clock, also the textclock would get 'squashed' and time would not be visible if, for example, a long command is typed into the prompt or there are too many icons in the systray.

By setting the align.horizontal layout to "none", this guarantees textclock will be in the middle, and that any overflow of right/left widgets do not affect it. Much neater behaviour. Also allows removing arbitray calculation for constraints based on workarea width.