Use a logical 'and' before removal command.
If the user doesn't have awesome's config installed to the given
directory, the first `mv` command will fail and the repository files
removed thereafter. To prevent this from happening, use the logical
`and` instead and the user has the opportunity to create the directory
and re-run (instead of having to re-clone).