Using string.format() on the net widget in powerarrow-dark.
As the net usage varies, so the numbers on the widget does. The problem is
that it makes the whole widget bar to go back and forth as the numbers change.
Adding a 'padding' number of zeroes reduces the number of 'moves' of the
widget row, and does not reduces the visibility of the widgets.